Migrate to ring buffers in mem chart
Use ring buffers instead of growing arrays to pre-allocate all the space we'll ever need.

Note:
We intentionally add some code duplication and TODOs for the period of migration.
This will get deleted later.

+++ src/base/channel.h
@@ -1,10 +1,9 @@
 #pragma once
 
-#include "base.h"
-
 #include <atomic>
+#include <cstdint>
 
-template <class T, size_t N> struct Channel {
+template <class T, uint32_t N> struct Channel {
   static constexpr uint32_t MASK = N - 1;
   std::atomic<uint32_t> head;
   std::atomic<uint32_t> tail;
